Login, logout, and sign-up features complete
We basically merged the login-logout functionality from http://cl.ly/0v0G1I273W3o
into our survey project.  We can now sign up, sign in, and sign out.

Basically we the login id from a unique e-mail to a unique username.
This is why you see a lot of lines changing the e-mail to the username.

We also updated some migrations that we had forgotten to add 's' to, which
meant the activerecord associations were not working right.

The rest of it is a straight merge from the aformentioned sinatra
skeleton.