Tested User.new() function
This tests activeRecord and ruby classes.

I expected User.new to create a new instance of the user object.   In my terms, the User class is a blueprint.  The instance is the house you build from that blueprint.

User inherits from the Activerecord class.
  In my terms, inheritance means that we are actually using the activeRecord blueprint.  ActiveRecord is a general blueprint, like the fact that a car has four wheels.  Then the User class is like a specific make and model of car.  Usually, when people talk about their car, they say things like my 'chevy' or my 'prius.'  So when you are referring to specific type of car, you refer to it by the more specific name, which is why in code we talk about User.new instead of Activerecord.User.new.

Activerecord is not actually a more general type of car. Instead, you know how it takes a whole sea of paperwork to keep your car running?  It takes a license and registration, and insurance, and records upon records, and you have to tell a whole bunch of people that you own a car and that you are maintaining it.  Wouldn't it be nice if you had a manager that took care of that stuff for you?  Activerecord does that.

When you build a car, you get to do all of the fun stuff of actually building it, and Activerecord takes care of keeping track of every car you build.

Now for the new function.  Using [.new](http://guides.rubyonrails.org/active_record_basics.html#create), as I understand it, creates a new instance without saving it to the database, like you would build a car without registering it with the police.

Note: railsguides are much better than (ActiveRecord/Base)[http://api.rubyonrails.org/classes/ActiveRecord/Base.html] documentation.