Tested four ways to initialize values when creating a new instance
As the [RailsGuides docs state,](http://guides.rubyonrails.org/active_record_basics.html#create) There are four ways to pass in parameters when you build a new object.

Let's talk about initializing parameters after the fact. Let's say you are building a house.  You can build the house. To finish the house, you need to bolt numbers on the sides of the buildings and ask the person who is about to move in what colors they want to paint it.  That is nice to know, but that information does not affect the pouring of the foundation and building the walls, so you can go head and do that. After that, the new homeowner can bolt on the letters and paint the walls herself.  That's what you are doing when you initialize parameters after the fact.

Now, let's talk about passing in parameters with a hash. In our house analogy, it would be like giving a builders a bag with the numbers inside and a bucket of paint. They can then go ahead and finish up the house for you. The hash is just a convenient way, the bag, to give the builders the nubmers and the paint. The hash is just one thing, whereas if you had to pass in the paint and the numbers individually, it would be like five things.

The thing is, you don't actually have to say that you're giving the builders a bag of stuff, because if you use the right notation, the bag is assumed. Thats why I can skip the curly braces.

Now, let's talk about passing in parameters with a block. I have no good analogy for that. I just know it works, although I don't know why you would use it.