Added a test to test bcrypt's password function in User
You know how some people store their passwords on a sticky note by their computer?  That's the equivalent of storing a password unencrypted in a database.  Anyone who can read it can get into the database.

Bcrypt is like taking that sticky and writing "You shall not pass!" on it when your password is Balrog. It doesn't store the password, it just stores something coded so that only you can get into the password.

Things I knew but forgot I knew: Include is for modules.  Bcrypt is a module.  If a class is a blueprint, a module is a bunch of standard tools that it would be convienent for the person living in the house. They're like appliences and furniture, sort of like stoves and refrigerators, or better yet, a mixer. Not everybody needs a mixer, but you can definitely mix-in a mixer in the houses that need one.

So bcrypt is a collection of things it would be really useful for our User to have. Specifically, it would be really convenient to not stick the password on a sticky. Instead, we ought to encoude the sticky.

See .question #1 in questions file for more questions about bcrypt.