Added a test to see if User.authenticate works
And it does.

Notes: User.authenticate is a class method, which makes sense, because we are trying to find out if a user exists in the database, and if so, if the password matches the things in the database.

A class method is like...lets see, if a class is the blueprints for a house, a class method is like the building inspector tasked with going around to all of the houses and seeing that they are built according to the blueprints.  A class method deals with every object ever created from those blueprints.