Added User password validation test
Added a test to see if the user gave a password that was at least six characters long.

Leassons learned: activerecord gives this fascinating...I'm not sure if it's a method or parameter. But it's called errors. So when activerecord is not successful in sticking something in the database, it still keeps the ruby object around, and fills it full of error messages so that a programmer has a fighting chance of figuring out what is wrong.  Activerecord errors is like an english teacher sending back a paper full of red ink and instructions to do it over.