Tested User-Surveys one to many association.
Found a bug. The User-Survey one-to-many association was not connected because the the column name was named users and not user in the database.

I deciced to 'fix' this by telling Activerecord, "You know that thing called a foreign key, the one you expected to be named 'user_id?' It's actually named users_id."  Activerecord relationships are a lot like actual relationships.  They don't work unless both people know that thye are in a relationship.  So I had to tell both the user and the survey models that they were connected, and they were connected with a key called users_id.